check-kernelpage: Add compatibility for python2 with urllib
With urllib2 the request module does not exist and urlretrieve is exposed
globally in the urllib namespace, while it is exposed in the request module
with python3. This commit fixes the issue by exporting the symbol urlretrieve
globally and removing the calls to the global ns, that is not portable between
two major version of python.

Signed-off-by: Romain Perier <romain.perier@gmail.com>