check-kernelpage: Add support for lzma with python2
The tarfile package can autodetect tar files extensions and select the right
program for uncompress it, like gzip, bzip2 or xz. The extensions .xz is
supported with python3 but not with python2. So instead of using tarfile.open()
directly, we define a function for extracting a xz archive that is used with
both version of python.

Moreover, lzma shipped with python3 and pyliblzma in python2, expose the same
module but use a different prototype for the open function. So fallback to
backports.lzma when python2 is being used to get exactly the same behaviours
with both version.

Signed-off-by: Romain Perier <romain.perier@gmail.com>