patch-kernel: Add support for config fragment per architecture
This adds a new option for the desired target architecture. Then the
default configuration to be used is found according to the kernel
version and the architecture. Moreover, the kernel is built out of the
source directory in order to share the same source between kernel
builders of the same branch and not collide together during the build.

Signed-off-by: Romain Perier <romain.perier@gmail.com>