check-kernelpage.py: Migration to kernel.org/releases.json
This updates the script to use releases.json. This is a JSON file hosted
on upstream that contains everything we need to detect if a new release
bug fix or just a release is available. We can know if this is an
incremental version, with the corresponding patch, or if this is a
full version instead and a lot of other metadata.

This commit updates the script to extract informations we need from this
JSON file, then it factorizes the code bit. Typically it makes the code
for incremental and not incremental version more generic. It also makes
the code for chmod simpler and faster by using os.chmod directly (this
typically avoids a lot of syscalls)

Signed-off-by: Romain Perier <romain.perier@gmail.com>