Remove experimental patches from patch list
Presently, there is some logic in ebuilds which selectively
applies patches depending on some environment parameters.
Until we have a mechanism in kernelCI to replicate this,
disable applying experimental patches for now.
Experimental patches may/may not be fully supported, but
these can be monitored separately from the main genpatches
set for the time being.
At some stage in the future, when we have finer-grained
application of patches in the CI system, these can probably
be added back in to the mix.

Signed-off-by: M. J. Everitt <m.j.everitt@iee.org>